Use new compiler options based on the ones that already exist
Summary:
# What?
Set the new options based on the ones that are currently in use.

`legacy_arrays_` replaces `no_use_hack_collections_`.
`hack_collections_`, which was the implicit default in the absence of `arrays_` and `no_use_hack_collections_`, is now explicit.
`arrays_` will eventually become the default and cease to exist as an explicit option.

# Why?

To deliver incremental changes without the need to change user specified compiler options, translate the compiler options specified as the input to the generator to the new options. More context below.

# Fixtures
Generated fixtures must not change.

# Context
`arraysets`, `no_use_hack_collections`, `stricttypes`, `array_migration`, `shape_arraykeys`, `const_collections` are all compiler options that control the generation of container fields. To make code generator simpler and easier to reason about, identify which of these can be removed/merged.

The new options based on the ones that are currently in use:

`legacy_arrays` replaces `no_use_hack_collections`.
`hack_collections`, which was the implicit default in the absence of `arrays` and `no_use_hack_collections`, is now explicit.
`arrays` will eventually become the default and cease to exist as an explicit option.

# The steps
The item in bold corresponds to the current diff.
1. **Use new compiler options based on the ones that already exist.**
1. Make arrays the default option.
1. Use the legacy arrays + hack collections logical equivalent of arrays.
1. Introduce hack collections wherever necessary.
1. Replace no use hack collections with legacy arrays.
1. Remove arraysets if arrays present.
1. Rename `no_use_hack_collections` compiler option to `legacy_arrays`.
1. Add `hack_collections` compiler option.
1. Remove references to the `arrays` compiler option.
1. Delete the `arrays` compiler option.

Reviewed By: yfeldblum

Differential Revision: D67545588

fbshipit-source-id: 909fa5e2875c1b32b256f47c454ae5c38662b559