Use the legacy arrays + hack collections logical equivalent of arrays.
Summary:
# What?
Replace `arrays_` with the equivalent logic based on `legacy_arrays_` and `hack_collections_`

# Why?
`arrays` are currently explicitly added by default. Make `arrays` the implicit default option, which will reduce the number of compiler options and simplify the logic in the process.

# Fixtures
Generated fixtures must not change.

# Context
`arraysets`, `no_use_hack_collections`, `stricttypes`, `array_migration`, `shape_arraykeys`, `const_collections` are all compiler options that control the generation of container fields. To make code generator simpler and easier to reason about, identify which of these can be removed/merged.

The new options based on the ones that are currently in use:

`legacy_arrays` replaces `no_use_hack_collections`.
`hack_collections`, which was the implicit default in the absence of `arrays` and `no_use_hack_collections`, is now explicit.
`arrays` will eventually become the default and cease to exist as an explicit option.

# The steps
The item in bold corresponds to the current diff.
1. Use new compiler options based on the ones that already exist.
1. Make arrays the default option.
1. **Use the legacy arrays + hack collections logical equivalent of arrays.**
1. Introduce hack collections wherever necessary.
1. Replace no use hack collections with legacy arrays.
1. Remove arraysets if arrays present.
1. Rename `no_use_hack_collections` compiler option to `legacy_arrays`.
1. Add `hack_collections` compiler option.
1. Remove references to the `arrays` compiler option.
1. Delete the `arrays` compiler option.

Reviewed By: rmakheja

Differential Revision: D67586337

fbshipit-source-id: a800c183f8814810dc8efed4c0ac04e15bdd394e