Add support for literals
Summary:
Take 2 of D67608814 that was reverted due to merge race causing build failure.

 ---

This diff adds support for literals as described in the spec.

These become useful with user-defined functions, which are introduced in future diffs.

The lexer already supports lexing literals so this was pretty easy to implement.

Reviewed By: iahs

Differential Revision: D67658832

fbshipit-source-id: af4ed5b3b9853e7eb9c3ecad8b3185b065076b88