Make const_collections imply hack_collections
Summary:
# What?
Generate an error if `hack_collections` appear along with `const_collections`.

# Why?
`const_collections` imply `hack_collections`. To eliminate unnecessary combinations of options, make it an error when both appear together.

# Fixtures
- Generated fixtures must not change.

# Context
`arraysets`, `no_use_hack_collections`, `stricttypes`, `array_migration`, `shape_arraykeys`, `const_collections` are all compiler options that control the generation of container fields. To make code generator simpler and easier to reason about, identify which of these can be removed/merged.

The new options based on the ones that are currently in use:

`legacy_arrays` replaces `no_use_hack_collections`.
`hack_collections`, which was the implicit default in the absence of `arrays` and `no_use_hack_collections`, is now explicit.
`arrays` will eventually become the default and cease to exist as an explicit option.
`const_collections` imply `hack_collections` and `hack_collections` must not appear explicitly in the presence of `const_collections`.

# The steps
The item in bold corresponds to the current diff.
1. Use new compiler options based on the ones that already exist.
1. Make arrays the default option.
1. Use the legacy arrays + hack collections logical equivalent of arrays.
1. Introduce hack collections wherever necessary.
1. Replace no use hack collections with legacy arrays.
1. Remove arraysets if arrays present.
1. Rename `no_use_hack_collections` compiler option to `legacy_arrays`.
1. Add `hack_collections` compiler option.
1. Remove references to the `arrays` compiler option.
1. Delete the `arrays` compiler option.
1. Remove `const_collections` if not `hack_collections`.
1. **Make `const_collections` imply `hack_collections`.**

Reviewed By: rmakheja

Differential Revision: D67718460

fbshipit-source-id: b982e7ae4430365871d5012a9aa62493ac579f69