Log dummy operations
Summary:
Previously we added logging to log deprecated operations, so that if there are no deprecated operations in production, we can remove the relevant code.

However, if there are no deprecated operations in scuba, it's hard to tell whether logging has a bug, or we really didn't have deprecated operations.

This diff will rule-out the former case by logging some dummy operations so that we know the logging path works.

Reviewed By: thedavekwon

Differential Revision: D67959504

fbshipit-source-id: 4e2439954b68ddfb83155e174876f77a0fd93778