Remove PrivacyUniverseUtils
Summary:
utilworst

This is a good illustration why the util class here is bad. It makes it appear like this is the canonical way to get the request universe, but really it isn't. It only works on thrift requests and even then it will be wrong if any inner zones are executing!

So instead I just kill the util wrapper and instead make the thrift context prop do the wrapping into the designator.

Reviewed By: gkunasag-fb

Differential Revision: D68153212

fbshipit-source-id: bdf0c718aaa55ed92075984a9608a32209e07f5e