Add abandonRead() to cursor readers
Summary:
Callers may wish to escape out of a nested read without performing the skipping to the end normally required to establish postconditions on the parent readers. That skipping can have significant CPU cost, and takes away from the benefit of CurSe when used to read prefixes of large structs.
The API enforces that it is called virally once a child has been abandoned.
There is no equivalent for writer because we never want to allow writing an incomplete buffer, whereas reading incompletely has no observable effects.

Reviewed By: yoney

Differential Revision: D68300807

fbshipit-source-id: 283509f773fcb9b4cba0a3d8ea0dd522fcc36caa