Remove parse_and_get_program
Summary:
`parse_and_get_program` is a legacy API that has a number of problems:

* It takes stringly-typed arguments, most of which are ignored.
* It may return an invalid AST (missing includes, unresolved types) which is not obvious at the call sites.
* It claims to not run mutators which doesn't make sense since mutation is an important part of semantic analysis and should be run e.g. for type inference.

It was virtually unused and most of its uses have already been migrated to `parse_ast` which doesn't have any of these problems and has an explicit opt in for behaviors such as ignoring missing includes that need special handling from consumers. Remove it together with migrating the last usage.

Reviewed By: createdbysk

Differential Revision: D68524407

fbshipit-source-id: 54705f0925522fbcc4f42d55acbc76051fe9f3e0