Introduce non-fatal parser errors
Summary:
This diff makes `report_error` *eventually* fail the parse (return `std::nullopt` instead of `ast::root`).

This allows the compiler to perform error-recovery — cases where the user intent is somewhat clear but expressed incorrectly.

The value is that instead of failing on the first error, we can collect multiple errors in a single parse and show them all to the user.

Reviewed By: rmakheja

Differential Revision: D68672819

fbshipit-source-id: 6e25e201e71c53ab23181ff887f8ef063c13df22