Add a capi conversion no-op method to thrift python structs/unions
Summary: There's a generated python capi method on thrift-py3 structs to initialize from thrift python types. It's a very simple fix to add a no-op method to the base thrift python StructOrUnion class to get these unit tests to start passing now. Since the method is meant to initialize a thrift python struct with a thrift python struct, all that we need to do is return the passed in object (after a basic instance check to make sure no python shenanigans are going on).

Reviewed By: ahilger

Differential Revision: D68867840

fbshipit-source-id: 817381af3586cf7586532c6f94af676cfc16b241