Make reflection work on thrift-py3 container flatnames when auto-migrate is enabled
Summary:
Reflection has to work on container flatnames in py3 because they're used recursively by the reflection library, so anything that expects to be able to recurse through a thrift python struct (eg. the py3 converter) will error out upon hitting a list.

This makes it so that the special py3 flatnames correctly generate reflection information, and also changes the converter to check for thrift python list types correctly.

Reviewed By: ahilger

Differential Revision: D68822077

fbshipit-source-id: 3844d11f476a902f230e1809d4d82de70ad6a1dc