Split ast::text into glyph and whitespace parts
Summary:
This diff splits up `ast::text` into whitespace and non-whitespace parts.

In a templating language like Whisker, whitespace is *significant* so it makes sense to capture it in the AST. Indeed, we already have this distinction implemented in the lexer — we just throw it away. You can see that the standalone line scanner already making decisions around `tok::text` and `tok::whitespace`.

So far we've gotten away with using `std::string` that *happened* to be whitespace in cases where we needed it. This change make it "safe" at a type-system level.

Doing this now because future diffs make use of whitespace vs non-whitespace text and we don't want to perform redundant (and expensive) checks on `ast::text`.

Reviewed By: vitaut

Differential Revision: D69089005

fbshipit-source-id: 77bf8ce2462f34c5281d8797acb1b5e37e2e6a03